fix(spa-editor): e2e AI provider seeding goes through encrypted Dexie repository (#608)
The seedAiProvider helper expected window.__ZUSTAND_STORES__.ai.addProvider,
but no such Zustand store exists — AI providers are persisted in Dexie and
the addProvider use case (application/ai/add-provider.ts) encrypts the
apiKey before put. The helper was throwing 'Missing window.__ZUSTAND_STORES__'
deterministically across all retries.

Route seeding through window.__KAIORD_E2E_SEED__.aiProvider, a dev-only seam
exposed in dexie-database.ts next to the existing __KAIORD_DB__ export. The
seeder uses the real createDexieAiProviderRepository so the apiKey is
encrypted at rest and decrypted on read by hooks/use-ai-providers-live, just
like a user-created provider.

- helpers/seed-stores.ts: removed (seedGarminCredentials had zero callers)
- helpers/seed-ai-provider.ts: new, calls window.__KAIORD_E2E_SEED__
- coaching-dialog-redesign.spec.ts: import path updated
